Rewrite message command methods to use entities
Since the Telegram-API considers "/foo-bar" to be a command entity
"/foo", we had choose a behaviour for the
CommandArguments() method.
The specification (https://core.telegram.org/bots#commands) mandates a
space after the command name, so we decided to skip the first character
following the command entity.
However, Telegram clients render "/foo-bar" as if "/foo" was the command
and "-bar" additional text, so this alternative should be remembered.